// src/lib/types.ts
export type ProjectDataType = {
	name: string;
	subtitle: string;
	description: string;
	role: string;
	technologies: string[];
	image: string;
	video: string;
};

export type ContactInfo = {
	email: string;
	discord: string;
	twitter: string;
	linkedin: string;
	socialmedia: string;
	form: string;
};